week-5/requirements.txt
fastapi==0.85.1
filelock==3.8.0
Pillow==9.2.0
pydantic==1.10.2
pytest-cov==4.0.0
pytest==7.1.3
streamlit==1.13.0
tqdm==4.64.1
typer==0.6.1
uvicorn==0.19.0
wandb==0.13.3
seldon-core==1.14.1
